refactor : 오목 체크 객체를 제거하고 오목 보드 자체에서 판단하도록 수정
