--- src/test/kotlin/domain/board/OmokCheckTest.kt
@@ -1,133 +1,0 @@
-package domain.board
-
-import domain.fixture.omokBoardFixture
-import omok.domain.board.OmokBoard
-import omok.domain.board.OmokColumn
-import omok.domain.board.OmokRow
-import omok.domain.board.StoneStatus
-import omok.domain.point.Point
-import omok.domain.rule.OmokCheck
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-
-class OmokCheckTest {
-    private lateinit var board: OmokBoard
-
-    @BeforeEach
-    fun setUp() {
-        board = omokBoardFixture()
-    }
-
-    /**
-     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
-     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   8 ●──●──●──●──X──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   7 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   6 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   5 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   4 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
-     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
-     * */
-    @Test
-    fun `오목 테스트1`() {
-        board.addStone(Point(OmokColumn.A, OmokRow.EIGHT, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.B, OmokRow.EIGHT, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.C, OmokRow.EIGHT, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.D, OmokRow.EIGHT, StoneStatus.BLACK))
-        val result = OmokCheck(board).isOmok(Point(OmokColumn.E, OmokRow.EIGHT, StoneStatus.BLACK))
-        assertThat(result).isTrue()
-    }
-
-    /**
-     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
-     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   8 ●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   7 ├──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   6 ├──┼──X──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   5 ├──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   4 ├──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
-     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
-     * */
-    @Test
-    fun `오목 테스트2`() {
-        board.addStone(Point(OmokColumn.A, OmokRow.EIGHT, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.B, OmokRow.SEVEN, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.D, OmokRow.FIVE, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.E, OmokRow.FOUR, StoneStatus.BLACK))
-        val result = OmokCheck(board).isOmok(Point(OmokColumn.C, OmokRow.SIX, StoneStatus.BLACK))
-        assertThat(result).isTrue()
-    }
-
-    /**
-     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
-     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  12 ├──┼──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  11 ├──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  10 ├──┼──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   9 ├──X──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   8 ●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   7 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   6 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   5 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   4 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
-     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
-     * */
-    @Test
-    fun `오목 테스트3`() {
-        board.addStone(Point(OmokColumn.A, OmokRow.EIGHT, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.C, OmokRow.TEN, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.D, OmokRow.ELEVEN, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.E, OmokRow.TWELVE, StoneStatus.BLACK))
-        val result = OmokCheck(board).isOmok(Point(OmokColumn.B, OmokRow.NINE, StoneStatus.BLACK))
-        assertThat(result).isTrue()
-    }
-
-    /**
-     *  15 ┌──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┬──┐
-     *  14 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  13 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  12 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  11 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *  10 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   9 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   8 ●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   7 ├──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   6 ├──┼──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   5 ├──┼──┼──●──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   4 ├──┼──┼──┼──X──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   3 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   2 ├──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┼──┤
-     *   1 └──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┴──┘
-     *     A  B  C  D  E  F  G  H  I  J  K  L  M  N  O
-     * */
-    @Test
-    fun `오목 테스트4`() {
-        board.addStone(Point(OmokColumn.A, OmokRow.EIGHT, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.B, OmokRow.SEVEN, StoneStatus.WHITE))
-        board.addStone(Point(OmokColumn.C, OmokRow.SIX, StoneStatus.BLACK))
-        board.addStone(Point(OmokColumn.D, OmokRow.FIVE, StoneStatus.BLACK))
-        val result = OmokCheck(board).isOmok(Point(OmokColumn.E, OmokRow.FOUR, StoneStatus.BLACK))
-        assertThat(result).isFalse()
-    }
-}
